Introducing SatisfactorySaveParser.ValueTypes namespace, ISerializableValue, basic value types
ISerializableValue is a base interface that represents a basic, standalone data type and its serialization format, without any of the SerializedProperty metadata (property name/index, encoded type name, etc). Value types generally implement the interface as ISerializableValue<T>, where T is the SerializableValue type itself, *not* the underlying data type (though for some records etc, these two might end up being the same). It's a little weirdly recursive, but it's also the best solution I've found for the limitations of C#'s type system.

ISerializableValue has two methods defined: Serialize(writer), which does what it says on the tin, and Parse(reader), which is an instance method that returns a new copy of its own class, and really ought to be static if C# would let me put that in the interface implementation. It doesn't, so instead you need to already have an instance of the SerializableValue type in question in order to parse a new one. C'est la vie.

This also includes some implementations of the basic value types, your ints and strings etc, and a couple compound types. The primitives in particular are defined as readonly structs with a single member and implicit conversions to and from the primitive type, which in my testing has functioned as an exact drop-in replacement for the base type itself in the JIT-compiled code - the disassembly reads as though the (e.g.) IntValue struct was just being handled as an int variable in comparisons and arithmetic operations, with the conversions to/from struct form getting completely optimized away. Neat.